Refactor indexables-route to indexable-action
